Move http request queue into `runtime-client` (#1745)
