+++ web-common/src/runtime-client/http-request-queue/HttpRequestQueue.spec.ts
@@ -4,8 +4,8 @@ import {
   runtimeServiceGetNumericHistogram,
 } from "@rilldata/web-common/runtime-client";
 import { httpRequestQueue } from "@rilldata/web-common/runtime-client/http-client";
-import { UrlExtractorRegex } from "@rilldata/web-local/lib/http-request-queue/HttpRequestQueue";
-import type { RequestQueueEntry } from "@rilldata/web-local/lib/http-request-queue/HttpRequestQueueTypes";
+import { UrlExtractorRegex } from "@rilldata/web-common/runtime-client/http-request-queue/HttpRequestQueue";
+import type { RequestQueueEntry } from "@rilldata/web-common/runtime-client/http-request-queue/HttpRequestQueueTypes";
 import { asyncWait, waitUntil } from "@rilldata/web-local/lib/util/waitUtils";
 import Mock = jest.Mock;
